requirements.txt
# Trading Strategy ML Requirements

# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Financial Data APIs
yfinance>=0.2.0
alpha-vantage>=2.3.0
pandas-datareader>=0.10.0

# Technical Analysis
TA-Lib>=0.4.25
# talib-binary>=0.4.19  # Not needed if TA-Lib is installed

# Machine Learning
tensorflow>=2.10.0
torch>=1.12.0
scikit-learn>=1.1.0
xgboost>=1.6.0
optuna>=3.0.0
lightgbm>=3.3.0

# Financial Analysis
backtrader>=1.9.76
zipline-reloaded>=2.2.0
arch>=5.3.0
quantlib>=1.29
pyfolio-reloaded>=0.9.5

# Database & Storage
psycopg2-binary>=2.9.0
redis>=4.3.0
sqlalchemy>=1.4.0
# timescaledb>=0.1.0  # Not available, use PostgreSQL with TimescaleDB extension

# Visualization
plotly>=5.10.0
matplotlib>=3.5.0
seaborn>=0.11.0
streamlit>=1.12.0
dash>=2.6.0

# Development & Testing
pytest>=7.1.0
pytest-cov>=3.0.0
black>=22.6.0
flake8>=5.0.0
mypy>=0.971

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
schedule>=1.2.0
tqdm>=4.64.0
joblib>=1.1.0

# Deployment
docker>=6.0.0
gunicorn>=20.1.0
uvicorn>=0.18.0

# Additional Financial Libraries
empyrical>=0.5.5
ffn>=0.3.7
pyfolio>=0.9.2
